requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
python-decouple==3.8
loguru==0.7.2
boto3==1.34.107
